Delete team helper from all files
