+++ lib/stat_tracker.rb
@@ -6,7 +6,6 @@ require_relative './game_teams'
 require_relative './modules/game_stats_module'
 require_relative './modules/league_stats_module'
 require_relative './modules/team_stats_module'
-require_relative './modules/team_stat_helper_module'
 require_relative './modules/season_stat_helper_module'
 require_relative './modules/season_stat_module'
 require_relative './modules/global_helper_module'
@@ -18,7 +17,6 @@ class StatTracker
   include GlobalHelpers
   include LeagueStats
   include TeamStats
-  include TeamStatHelpers
   include SeasonStat
   include SeasonStatHelper
